Allow external (re)initialization of the module.
When a graph is added dynamically through JavaScript (e.g. by an AJAX call), the pygal-toggle-graphs module has to be re-initialized by calling initPygalToggleGraphs(). Otherwise, toggling will not work. This is because the module sets on-click actions to all graphs that are present at the time it is load. When a graph is added later, the on-click actions are not set.